feat: form preview
