+++ src/features/listings/components/Submission/SubmissionDrawer.tsx
@@ -1,32 +1,17 @@
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useQueryClient } from '@tanstack/react-query';
-import { type ClassValue } from 'clsx';
-import { Check, ChevronDown, X } from 'lucide-react';
+import { X } from 'lucide-react';
 import { useRouter } from 'next/router';
 import { usePostHog } from 'posthog-js/react';
 import { type JSX, useEffect, useState } from 'react';
-import {
-  type Control,
-  useForm,
-  type UseFormReturn,
-  useWatch,
-} from 'react-hook-form';
+import { useForm, type UseFormReturn, useWatch } from 'react-hook-form';
 import { toast } from 'sonner';
 import { type z } from 'zod';
 
-import { RichEditor } from '@/components/shared/RichEditor';
-import { MinimalTiptapEditor } from '@/components/tiptap';
+import { EligibilityQuestions } from '@/components/eligibility/EligibilityQuestions';
 import { Alert, AlertDescription } from '@/components/ui/alert';
 import { Button } from '@/components/ui/button';
 import { Checkbox } from '@/components/ui/checkbox';
-import {
-  Command,
-  CommandEmpty,
-  CommandGroup,
-  CommandInput,
-  CommandItem,
-  CommandList,
-} from '@/components/ui/command';
 import {
   Form,
   FormControl,
@@ -34,25 +19,10 @@ import {
   FormField,
   FormItem,
   FormLabel,
-  FormMessage,
 } from '@/components/ui/form';
-import { FormFieldWrapper } from '@/components/ui/form-field-wrapper';
 import { Input } from '@/components/ui/input';
 import { KycComponent } from '@/components/ui/KycComponent';
-import {
-  Popover,
-  PopoverContent,
-  PopoverTrigger,
-} from '@/components/ui/popover';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
 import { SideDrawer, SideDrawerContent } from '@/components/ui/side-drawer';
-import { WalletConnectField } from '@/components/ui/wallet-connect-field';
 import { CHAIN_NAME } from '@/constants/project';
 import { tokenList } from '@/constants/tokenList';
 import { type SubmissionWithUser } from '@/interface/submission';
@@ -61,9 +31,7 @@ import { useUser } from '@/store/user';
 import { cn } from '@/utils/cn';
 
 import { AuthWrapper } from '@/features/auth/components/AuthWrapper';
-import { InfoBox } from '@/features/sponsor-dashboard/components/InfoBox';
 
-import { walletFieldListings } from '../../constants';
 import { submissionCountQuery } from '../../queries/submission-count';
 import { listingSubmissionsQuery } from '../../queries/submissions';
 import { userSubmissionQuery } from '../../queries/user-submission-status';
@@ -109,7 +77,6 @@ export const SubmissionDrawer = ({
   const queryClient = useQueryClient();
   const isProject = type === 'project';
   const isHackathon = type === 'hackathon';
-  const isSponsorship = type === 'sponsorship';
   const [isLoading, setIsLoading] = useState(false);
   const [isTOSModalOpen, setIsTOSModalOpen] = useState(false);
   const [termsAccepted, setTermsAccepted] = useState(false);
@@ -134,11 +101,6 @@ export const SubmissionDrawer = ({
   const formPublicKey = useWatch({
     control: form.control,
     name: 'publicKey',
-  });
-
-  const tokenSelected = useWatch({
-    control: form.control,
-    name: 'token',
   });
 
   const posthog = usePostHog();
@@ -352,350 +314,66 @@ export const SubmissionDrawer = ({
                 </div>
                 <div>
                   <div className="mb-5 flex flex-col gap-4">
-                    {!isProject &&
-                      !isSponsorship &&
-                      !walletFieldListings.includes(id!) && (
-                        <>
-                          <FormField
-                            control={form.control}
-                            name={'link'}
-                            render={({ field }) => (
-                              <FormItem className={cn('flex flex-col gap-2')}>
-                                <div>
-                                  <FormLabel isRequired>
-                                    Link to Your Submission
-                                  </FormLabel>
-                                  <FormDescription>
-                                    Make sure this link is accessible by
-                                    everyone!
-                                  </FormDescription>
-                                </div>
-                                <div>
-                                  <FormControl>
-                                    <div className="flex">
-                                      <div className="flex items-center gap-1 rounded-l-md border border-r-0 border-input bg-muted px-2 shadow-sm">
-                                        <p className="text-sm font-medium text-slate-500">
-                                          https://
-                                        </p>
-                                      </div>
-                                      <Input
-                                        {...field}
-                                        maxLength={500}
-                                        placeholder="Add a link"
-                                        className="rounded-l-none"
-                                        autoComplete="off"
-                                      />
-                                    </div>
-                                  </FormControl>
-
-                                  <FormMessage className="pt-1" />
-                                </div>
-                              </FormItem>
-                            )}
-                          />
-                          <FormField
-                            control={form.control}
-                            name={'tweet'}
-                            render={({ field }) => (
-                              <FormItem className={cn('flex flex-col gap-2')}>
-                                <div>
-                                  <FormLabel>Tweet Link</FormLabel>
-                                  <FormDescription>
-                                    This helps sponsors discover (and maybe
-                                    repost) your work on Twitter! If this
-                                    submission is for a Twitter thread bounty,
-                                    you can ignore this field.
-                                  </FormDescription>
-                                </div>
-                                <div>
-                                  <FormControl>
-                                    <div className="flex">
-                                      <div className="flex items-center gap-1 rounded-l-md border border-r-0 border-input bg-muted px-2 shadow-sm">
-                                        <p className="text-sm font-medium text-slate-500">
-                                          https://
-                                        </p>
-                                      </div>
-                                      <Input
-                                        {...field}
-                                        maxLength={500}
-                                        placeholder="Add a tweet's link"
-                                        className="rounded-l-none"
-                                        autoComplete="off"
-                                      />
-                                    </div>
-                                  </FormControl>
-
-                                  <FormMessage className="pt-1" />
-                                </div>
-                              </FormItem>
-                            )}
-                          />
-                        </>
-                      )}
-                    {eligibility?.map((e, index) => {
-                      if (
-                        walletFieldListings.includes(id!) &&
-                        e.question === `Connect Your ${CHAIN_NAME} Wallet`
-                      ) {
-                        return (
-                          <WalletConnectField
-                            key={`${index}-${e.order}`}
-                            control={form.control}
-                            name={`eligibilityAnswers.${index}.answer`}
-                            label={e.question}
-                            isRequired
-                            description="Connect your wallet to verify ownership. This is mandatory for this bounty."
-                          />
-                        );
-                      }
-                      if (e.type === 'checkbox') {
-                        return (
-                          <FormField
-                            key={`${index}-${e.order}`}
-                            control={form.control}
-                            name={`eligibilityAnswers.${index}.answer`}
-                            render={({ field }) => (
-                              <FormItem className={cn('flex flex-col gap-1')}>
-                                <div className="flex items-center gap-2">
-                                  <FormControl>
-                                    <Checkbox
-                                      checked={field.value === 'true'}
-                                      onCheckedChange={(checked) => {
-                                        field.onChange(
-                                          checked ? 'true' : 'false',
-                                        );
-                                      }}
-                                    />
-                                  </FormControl>
-                                  <InfoBox
-                                    content={e.question}
-                                    className="mb-0"
-                                    contentClassName={cn(
-                                      '[&_p]:!text-[0.9rem] [&_a]:!text-[0.9rem] flex items-center',
-                                      e.optional !== true &&
-                                        "after:content-['*'] after:text-red-500 after:ml-1",
-                                    )}
-                                    isHtml={true}
-                                  />
-                                </div>
-                                <FormMessage />
-                              </FormItem>
-                            )}
-                          />
-                        );
-                      } else if (e.type === 'select') {
-                        return (
-                          <FormField
-                            key={`${index}-${e.order}`}
-                            control={form.control}
-                            name={`eligibilityAnswers.${index}.answer`}
-                            render={({ field }) => (
-                              <FormItem className={cn('flex flex-col gap-2')}>
-                                <div>
-                                  <FormLabel isRequired={e.optional !== true}>
-                                    {e.question}
-                                  </FormLabel>
-                                  {e.description && (
-                                    <FormDescription className="whitespace-pre-wrap text-wrap">
-                                      {e.description}
-                                    </FormDescription>
-                                  )}
-                                </div>
-                                <div>
-                                  <FormControl>
-                                    <Select
-                                      value={field.value}
-                                      onValueChange={field.onChange}
-                                    >
-                                      <SelectTrigger>
-                                        <SelectValue placeholder="Select an option..." />
-                                      </SelectTrigger>
-                                      <SelectContent>
-                                        {e.variants?.map(
-                                          (
-                                            variant: string,
-                                            variantIndex: number,
-                                          ) => (
-                                            <SelectItem
-                                              key={variantIndex}
-                                              value={variant}
-                                            >
-                                              {variant}
-                                            </SelectItem>
-                                          ),
-                                        )}
-                                      </SelectContent>
-                                    </Select>
-                                  </FormControl>
-                                  <FormMessage className="pt-1" />
-                                </div>
-                              </FormItem>
-                            )}
-                          />
-                        );
-                      }
-
-                      return (
-                        <FormField
-                          key={`${index}-${e.order}`}
-                          control={form.control}
-                          name={`eligibilityAnswers.${index}.answer`}
-                          render={({ field }) => (
-                            <FormItem className={cn('flex flex-col gap-2')}>
-                              <div>
-                                <FormLabel isRequired={e.optional !== true}>
-                                  {e.question}
-                                </FormLabel>
-                                {e.description && (
-                                  <FormDescription className="whitespace-pre-wrap text-wrap">
-                                    {e.description}
-                                  </FormDescription>
-                                )}
-                              </div>
-                              <div>
-                                <FormControl>
-                                  {e.isLink || e.type === 'link' ? (
-                                    <div className="flex">
-                                      <div className="flex items-center gap-1 rounded-l-md border border-r-0 border-input bg-muted px-2 shadow-sm">
-                                        <p className="text-sm font-medium text-slate-500">
-                                          https://
-                                        </p>
-                                      </div>
-                                      <Input
-                                        {...field}
-                                        placeholder="Add a link..."
-                                        className="rounded-l-none"
-                                        autoComplete="off"
-                                      />
-                                    </div>
-                                  ) : e.type === 'paragraph' ? (
-                                    <div className="flex rounded-md border shadow-sm ring-primary has-[:focus]:ring-1">
-                                      <MinimalTiptapEditor
-                                        key={`${field.name}-${editFetched ? id : ''}`}
-                                        {...field}
-                                        value={field.value || ''}
-                                        immediatelyRender={false}
-                                        className="min-h-[30vh] w-full border-0 text-sm"
-                                        editorContentClassName="p-4 px-2 h-full"
-                                        output="html"
-                                        placeholder="Type your description here..."
-                                        editable={true}
-                                        editorClassName="focus:outline-none"
-                                        imageSetting={{
-                                          folderName:
-                                            'listing-eligibility-answer',
-                                          type: 'custom-question',
-                                        }}
-                                        toolbarClassName="sticky top-0 rounded-t-md bg-white z-[75] w-full overflow-x-hidden"
-                                      />
-                                    </div>
-                                  ) : (
-                                    <RichEditor
-                                      {...field}
-                                      id={`eligibilityAnswers.${index}.answer`}
-                                      value={field.value || ''}
-                                      error={false}
-                                      placeholder={'Write something...'}
-                                    />
-                                  )}
-                                </FormControl>
-                                <FormMessage className="pt-1" />
-                              </div>
-                            </FormItem>
-                          )}
-                        />
-                      );
-                    })}
-
-                    {compensationType !== 'fixed' && (
-                      <FormFieldWrapper
-                        control={form.control}
-                        name="ask"
-                        label="Total Amount"
-                        description={
-                          token !== 'Any'
-                            ? "What's the compensation you require to complete this fully?"
-                            : 'Enter the exact amount you are seeking in US dollars.'
-                        }
-                        isRequired
-                        isTokenInput
-                        token={token}
-                      />
-                    )}
-
-                    {token === 'Any' && <TokenSelect control={form.control} />}
-                    {token === 'Any' && tokenSelected === 'Other' && (
-                      <FormFieldWrapper
-                        control={form.control}
-                        name="otherTokenDetails"
-                        isRequired={tokenSelected === 'Other'}
-                        label="Payment details"
-                        isRichEditor
-                        description="What's your preferred way to receive a payment ?"
-                        richEditorPlaceholder="I want to receive a payment in..."
-                      />
-                    )}
-
-                    <FormFieldWrapper
+                    <EligibilityQuestions
+                      questions={eligibility}
                       control={form.control}
-                      name="otherInfo"
-                      label="Anything Else?"
-                      description="If you have any other links or information you'd like to share with us, please add them here!"
-                      isRichEditor
-                      richEditorPlaceholder="Add info or link"
+                      listingId={id ?? null}
+                      editFetched={editFetched}
+                      compensationType={compensationType ?? null}
+                      token={token ?? null}
+                      listingType={type ?? null}
                     />
-                    {!walletFieldListings.includes(id!) && (
-                      <FormField
-                        control={form.control}
-                        name="publicKey"
-                        render={({ field }) => (
-                          <FormItem className="flex w-full flex-col gap-2">
-                            <div>
-                              <FormLabel isRequired={!user?.publicKey}>
-                                Your {CHAIN_NAME} Wallet Address
-                              </FormLabel>
-                              <FormDescription>
-                                {!!user?.publicKey ? (
-                                  <>
-                                    This is where you will receive your payment
-                                    if your submission is approved. If you want
-                                    to edit it,{' '}
-                                    <a
-                                      href={`/t/${user?.username}/edit`}
-                                      className="text-blue-600 underline hover:text-blue-700"
-                                      target="_blank"
-                                      rel="noopener noreferrer"
-                                    >
-                                      click here
-                                    </a>
-                                  </>
-                                ) : (
-                                  <>
-                                    This wallet address will be linked to your
-                                    profile and you will receive your rewards
-                                    here if you win.
-                                  </>
+
+                    <FormField
+                      control={form.control}
+                      name="publicKey"
+                      render={({ field }) => (
+                        <FormItem className="flex w-full flex-col gap-2">
+                          <div>
+                            <FormLabel isRequired={!user?.publicKey}>
+                              Your {CHAIN_NAME} Wallet Address
+                            </FormLabel>
+                            <FormDescription>
+                              {!!user?.publicKey ? (
+                                <>
+                                  This is where you will receive your payment if
+                                  your submission is approved. If you want to
+                                  edit it,{' '}
+                                  <a
+                                    href={`/t/${user?.username}/edit`}
+                                    className="text-blue-600 underline hover:text-blue-700"
+                                    target="_blank"
+                                    rel="noopener noreferrer"
+                                  >
+                                    click here
+                                  </a>
+                                </>
+                              ) : (
+                                <>
+                                  This wallet address will be linked to your
+                                  profile and you will receive your rewards here
+                                  if you win.
+                                </>
+                              )}
+                            </FormDescription>
+                          </div>
+                          <FormControl>
+                            <div className="flex flex-col gap-2">
+                              <Input
+                                className={cn(
+                                  !!user?.publicKey &&
+                                    'cursor-not-allowed text-slate-600 opacity-80',
                                 )}
-                              </FormDescription>
+                                placeholder={`Add your ${CHAIN_NAME} wallet address`}
+                                readOnly={!!user?.publicKey}
+                                {...(!!user?.publicKey ? {} : field)}
+                                value={user?.publicKey || field.value}
+                              />
                             </div>
-                            <FormControl>
-                              <div className="flex flex-col gap-2">
-                                <Input
-                                  className={cn(
-                                    !!user?.publicKey &&
-                                      'cursor-not-allowed text-slate-600 opacity-80',
-                                  )}
-                                  placeholder={`Add your ${CHAIN_NAME} wallet address`}
-                                  readOnly={!!user?.publicKey}
-                                  {...(!!user?.publicKey ? {} : field)}
-                                  value={user?.publicKey || field.value}
-                                />
-                              </div>
-                            </FormControl>
-                          </FormItem>
-                        )}
-                      />
-                    )}
+                          </FormControl>
+                        </FormItem>
+                      )}
+                    />
                     <KycComponent
                       address={formPublicKey ?? user?.publicKey ?? ''}
                       listingSponsorId={listing.sponsorId}
@@ -795,175 +473,3 @@ export const SubmissionDrawer = ({
     </SideDrawer>
   );
 };
-
-interface TokenSelectProps {
-  control: Control<FormData>;
-}
-
-export function TokenSelect({ control }: TokenSelectProps) {
-  return (
-    <FormField
-      name="token"
-      control={control}
-      render={({ field }) => (
-        <FormItem className="gap-2">
-          <div>
-            <FormLabel>Currency</FormLabel>
-            <FormDescription>
-              Select your preferred currency for receiving funds. The exchange
-              rate will be the closing rate at the day of the invoice.
-            </FormDescription>
-          </div>
-          <Popover>
-            <PopoverTrigger asChild>
-              <FormControl>
-                <Button
-                  variant="outline"
-                  role="combobox"
-                  className={cn(
-                    'w-full justify-between',
-                    !field.value && 'text-muted-foreground',
-                  )}
-                >
-                  {field.value ? (
-                    <TokenLabel
-                      control={control}
-                      showIcon
-                      showSymbol
-                      classNames={{
-                        symbol: 'text-slate-900',
-                        postfix: 'text-slate-900',
-                      }}
-                    />
-                  ) : (
-                    'Select Token'
-                  )}
-                  <ChevronDown className="opacity-50" />
-                </Button>
-              </FormControl>
-            </PopoverTrigger>
-            <PopoverContent className="w-[33rem] p-0">
-              <Command>
-                <CommandInput placeholder="Search token..." className="h-9" />
-                <CommandList>
-                  <CommandEmpty>No Token found.</CommandEmpty>
-                  <CommandGroup>
-                    {tokenList
-                      .filter((token) => token.tokenSymbol !== 'Any')
-                      .map((token) => (
-                        <CommandItem
-                          value={token.tokenName}
-                          key={token.tokenSymbol}
-                          onSelect={() => {
-                            field.onChange(token.tokenSymbol);
-                          }}
-                        >
-                          <TokenLabel
-                            control={control}
-                            token={token}
-                            showIcon
-                            showName
-                          />
-                          <Check
-                            className={cn(
-                              'ml-auto',
-                              token.tokenSymbol === field.value
-                                ? 'opacity-100'
-                                : 'opacity-0',
-                            )}
-                          />
-                        </CommandItem>
-                      ))}
-                  </CommandGroup>
-                </CommandList>
-              </Command>
-            </PopoverContent>
-          </Popover>
-          <FormMessage />
-        </FormItem>
-      )}
-    />
-  );
-}
-
-interface TokenLabelProps {
-  symbol?: string;
-  className?: ClassValue;
-  showIcon?: boolean;
-  showSymbol?: boolean;
-  showName?: boolean;
-  postfix?: string;
-  amount?: number | null;
-  token?: (typeof tokenList)[0];
-  classNames?: {
-    icon?: ClassValue;
-    symbol?: ClassValue;
-    amount?: ClassValue;
-    postfix?: ClassValue;
-    name?: ClassValue;
-  };
-  control: Control<FormData>;
-  formatter?: (amount: number) => string;
-}
-
-const defaultFormatter = (amount: number) =>
-  new Intl.NumberFormat('en-US', {
-    currency: 'USD',
-    maximumFractionDigits: 2,
-  }).format(amount);
-
-export function TokenLabel({
-  symbol,
-  className,
-  showIcon = true,
-  showSymbol = false,
-  showName = false,
-  postfix,
-  amount,
-  classNames,
-  token: preToken,
-  formatter = defaultFormatter,
-  control,
-}: TokenLabelProps) {
-  const formToken = useWatch({
-    control: control,
-    name: 'token',
-  });
-
-  const searchSymbol = symbol || formToken;
-  const token =
-    preToken || tokenList.find((token) => token.tokenSymbol === searchSymbol);
-
-  if (!token) return null;
-  return (
-    <span className={cn('flex w-max items-center', className)}>
-      {showIcon && (
-        <img
-          src={token.icon}
-          alt={token.tokenSymbol}
-          className={cn('mr-1 block h-4 w-4', classNames?.icon)}
-        />
-      )}
-      {typeof amount === 'number' && !isNaN(amount) && (
-        <span className={cn('ml-2 text-sm', classNames?.amount)}>
-          {formatter(amount)}
-        </span>
-      )}
-      {showName && (
-        <span className={cn('ml-2 text-sm', classNames?.symbol)}>
-          {token.tokenName}
-        </span>
-      )}
-      {showSymbol && (
-        <span className={cn('ml-2 text-sm', classNames?.symbol)}>
-          {token.tokenSymbol}
-        </span>
-      )}
-      {postfix && (
-        <span className={cn('ml-1 text-sm', classNames?.postfix)}>
-          {postfix}
-        </span>
-      )}
-    </span>
-  );
-}
